refactor autocomplete system
